==> notebook/1. Review_Translation.ipynb <==
# 1. Read Dataset

import pandas as pd
import os

def process_row(row):
    """
    Processes a single row from the CSV file. If the row has more than 4 elements, it keeps the first, second, and last elements,
    and merges the middle elements into a single string.

    Parameters:
    - row (list of str): The row to be processed, represented as a list of string elements.

    Returns:
    - list of str: A processed row with the first, second, and last elements, and a merged string of any middle elements.
      If the original row had 4 or fewer elements, it is returned unchanged.
    """
    if len(row) > 4:
        first, second, *middle, last = row
        merged_middle = ' '.join(middle)
        return [first, second, merged_middle, last]
    return row

def create_dataframe_from_csv(file_path):
    """
    Reads a CSV file from the given file path and creates a pandas DataFrame. Each row in the CSV is processed to merge
    middle elements if the row contains more than 4 elements.

    Parameters:
    - file_path (str): The path to the CSV file to be read.

    Returns:
    - pd.DataFrame: A pandas DataFrame created from the processed CSV data. The first row of the CSV is used as column headers.
    """
    data = []
    with open(file_path, 'r', encoding='utf-8') as file:
        for line in file:
            row = [item.strip().strip('"') for item in line.split(',')]
            processed_row = process_row(row)
            data.append(processed_row)

    headers = data[0]
    df_data = data[1:]
    df = pd.DataFrame(df_data, columns=headers)
    return df

def preprocess_dataframe(df):
    """
    Preprocesses the given DataFrame by dropping the 'id' column and removing duplicates based on 'app' and 'content' columns.

    Parameters:
    - df (pd.DataFrame): The input DataFrame to preprocess.

    Returns:
    - pd.DataFrame: The preprocessed DataFrame, with the 'id' column dropped and duplicates removed.
    """
    df.drop(columns=['id'], inplace=True)
    df.drop_duplicates(subset=['app', 'content'], inplace=True)
    return df


base_dir = '../dataset/phase 1/'  
file_path_1 = os.path.join(base_dir, 'google_app_reviews_1.csv')
file_path_2 = os.path.join(base_dir, 'google_app_reviews_2.csv')

# Creating DataFrames from CSV files
df_1 = create_dataframe_from_csv(file_path_1)
df_2 = create_dataframe_from_csv(file_path_2)

# Preprocessing DataFrames
df_1 = preprocess_dataframe(df_1)
df_2 = preprocess_dataframe(df_2)

# Concatenating preprocessed DataFrames
df = pd.concat([df_1, df_2], ignore_index=True).reset_index(drop=True)


# 2. User Review Translation
## 2.1. Google Play Store

import pandas as pd
import time
import warnings
from transformers import MarianMTModel, MarianTokenizer

def translate_content_and_save(df, app_names, model_name, output_path, batch_size=10):
    """
    Translates the content of specific rows in a DataFrame where the 'app' column matches any of the given app_names,
    using the specified MarianMT model. It then saves the updated DataFrame to a CSV file at the specified output_path.

    Parameters:
    - df (pd.DataFrame): The input DataFrame containing the data to be translated.
    - app_names (list of str): The names of the apps for which the 'content' column will be translated.
    - model_name (str): The MarianMT model name to be used for translation.
    - output_path (str): The file path where the updated DataFrame should be saved as a CSV.
    - batch_size (int, optional): The number of sentences to translate before printing progress. Defaults to 10.

    Returns:
    - None: This function does not return a value but saves the updated DataFrame to a CSV file.

    This function filters the DataFrame for rows corresponding to any of the specified app_names, translates the 'content'
    column of those rows, and then merges the translated content back into the original DataFrame. Finally, it saves
    the updated DataFrame to a CSV file at the specified output_path. Progress is printed every batch_size translations.
    """
    warnings.filterwarnings('ignore')
    translated_dfs = []

    for app_name in app_names:
        # Filter DataFrame for the specified app
        df_app_specific = df[df.app == app_name]
        print(df_app_specific.head())

        # Initialize translation function
        def translate(text, model_name=model_name):
            try:
                tokenizer = MarianTokenizer.from_pretrained(model_name)
                model = MarianMTModel.from_pretrained(model_name)
                tokens = tokenizer(text, return_tensors="pt", truncation=True, max_length=512)
                translated = model.generate(**tokens)
                translated_text = tokenizer.decode(translated[0], skip_special_tokens=True)
                return translated_text
            except Exception as e:
                print(f"Error in translation: {e}")
                return 'Translation failed'

        # Perform translation
        start_time = time.time()
        translated_text = []
        for index, content in enumerate(df_app_specific.content):
            translated_text.append(translate(content))
            if (index + 1) % batch_size == 0:
                duration = time.time() - start_time
                print(f"Successfully translated {index + 1} sentences for app '{app_name}' in {duration:.2f} seconds")

        # Update DataFrame with translated content
        df_app_specific = df_app_specific.reset_index(drop=True)
        df_app_specific['content'] = pd.Series(translated_text)

        translated_dfs.append(df_app_specific)

    # Concatenate translated DataFrames with the rest of the original DataFrame
    df_non_translated = df[~df.app.isin(app_names)]
    df_updated = pd.concat([df_non_translated] + translated_dfs, ignore_index=True).reset_index(drop=True)

    # Save to CSV
    print(df_updated.head())
    #df_updated.to_csv(output_path, index=False)
    print(f"Dataframe has been saved into {output_path}!")
    return df_updated


### 2.1.1. Polish to English

App: 112 BE

# Calling the parameters
save_dir = '../dataset/phase 2/' 
app_name = ['Alarm112']
model_name = 'Helsinki-NLP/opus-mt-pl-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df, app_name, model_name, output_path)

### 2.1.2. Dutch to English

App: 112 BE

# Calling the parameters
app_name = ['112 BE']
model_name = 'Helsinki-NLP/opus-mt-nl-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.3. German to English

App: DEC112 2.0

# Calling the parameters
app_name = ['DEC112 2.0']
model_name = 'Helsinki-NLP/opus-mt-de-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.4. Georgian to English

App: 112 Georgia

# Calling the parameters
app_name = ['112 Georgia']
model_name = 'Helsinki-NLP/opus-mt-ka-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.5. Spanish to English

App: SAC Argentina

# Calling the parameters
app_name = ['SAC Argentina']
model_name = 'Helsinki-NLP/opus-mt-es-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.6. Bulgarian to English

App: help is nearby!

# Calling the parameters
app_name = ['help is nearby!']
model_name = 'Helsinki-NLP/opus-mt-bg-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.7. Czech to English

App: Zachranka

# Calling the parameters
app_name = ['Zachranka']
model_name = 'Helsinki-NLP/opus-mt-cs-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.8. Hungarian to English

App: EletMento

# Calling the parameters
app_name = ['EletMento']
model_name = 'Helsinki-NLP/opus-mt-hu-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.9. Bahasa to English

App: Jakarta Aman

# Calling the parameters
app_name = ['Jakarta Aman']
model_name = 'Helsinki-NLP/opus-mt-id-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.10. Japanesse to English

App: Yurekuru Call and National evacuation center guide

# Calling the parameters
app_name = ['Yurekuru Call', 'National evacuation center guide']
model_name = 'Helsinki-NLP/opus-mt-ja-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.1.11. Mongolian to English

App: Anhaar

# Calling the parameters
app_name = ['Anhaar']
model_name = 'Helsinki-NLP/opus-mt-mul-en'
output_path = os.path.join(save_dir, 'df_translated_google_play.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

## 2.2. Apple App Store

file_path = os.path.join(save_dir, 'apple_app_reviews.csv')
df = pd.read_csv(file_path)

### 2.2.1. Polish to English

App: 112 BE

# Calling the parameters
app_name = ['Alarm112']
model_name = 'Helsinki-NLP/opus-mt-pl-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df, app_name, model_name, output_path)

### 2.2.2. Dutch to English

App: 112 BE

# Calling the parameters
app_name = ['112 BE']
model_name = 'Helsinki-NLP/opus-mt-nl-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.3. Bulgarian to English

App: 112 Bulgaria

# Calling the parameters
app_name = ['112 Bulgaria']
model_name = 'Helsinki-NLP/opus-mt-bg-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.4. Japanesse to English

App: PREP

# Calling the parameters
app_name = ['PREP']
model_name = 'Helsinki-NLP/opus-mt-ja-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.5. Norwegian to English

App: Hjelp 113

#### 2.2.5.1 Norwegian to Dutch

# Calling the parameters
app_name = ['Hjelp 113']
model_name = 'Helsinki-NLP/opus-mt-no-nl'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

#### 2.2.5.2. Dutch to English

# Calling the parameters
app_name = ['Hjelp 113']
model_name = 'Helsinki-NLP/opus-mt-nl-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.6. Swedish to English

App: SOS Alarm

# Calling the parameters
app_name = ['SOS Alarm']
model_name = 'Helsinki-NLP/opus-mt-sv-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.7. Swiss to English

App: Alertswiss

#### 2.2.7.1. German to English

# Calling the parameters
app_name = ['Alertswiss']
model_name = 'Helsinki-NLP/opus-mt-de-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

#### 2.2.7.2. French to English

# Calling the parameters
app_name = ['Alertswiss']
model_name = 'Helsinki-NLP/opus-mt-fr-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.8. Turkish to English

App: Afad Acil Cagri

# Calling the parameters
app_name = ['Afad Acil Cagri']
model_name = 'Helsinki-NLP/opus-mt-tr-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)

### 2.2.9. Spanish to English

App: Earthquake and My Earthquake Alerts & Feed

# Calling the parameters
app_name = ['Earthquake', 'My Earthquake Alerts & Feed']
model_name = 'Helsinki-NLP/opus-mt-es-en'
output_path = os.path.join(save_dir, 'df_translated_app_store.csv')

# Calling the function
df_updated = translate_content_and_save(df_updated, app_name, model_name, output_path)
